Config:Setup linters and file for github actions
